fix(reload): restore reconfigure contract
Keep runtime configuration wrapping lifecycle-internal, share modal test
plan setup, and remove unrelated pending-target grammar extraction.
